Integrate pngfilters in png.py.
